# docs/swagger.yaml
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/user/create:
    post:
      description: 创建一个用户，并附带默认配置，默认密码为dephy@2025.com
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 用户电话号码
        in: body
        name: phone
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 创建用户失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 创建用户
      tags:
      - 用户管理
  /api/v1/user/delete:
    get:
      description: 根据用户uuid参数删除当前用户
      parameters:
      - description: 用户uuid
        in: query
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 删除用户失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 删除用户
      tags:
      - 用户管理
  /api/v1/user/list:
    get:
      description: 获取所有矿机数据
      parameters:
      - description: 分页查询第几页
        in: query
        name: page
        required: true
        type: string
      - description: 每页限制查询的数据数量
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 获取矿机列表失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 获取矿机列表
      tags:
      - 机器管理
  /api/v1/user/login:
    post:
      description: 用户登录接口
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 用户电话号码
        in: query
        name: phone
        required: true
        type: string
      - description: 用户密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 登录失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 登录接口
      tags:
      - 用户管理
  /api/v1/user/logout:
    post:
      description: 当前用户登出操作
      parameters:
      - description: 用户token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 用户登出失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 用户登出
      tags:
      - 用户管理
  /api/v1/user/update:
    post:
      description: 更新一个用户的所有信息，根据uuid搜索用户
      parameters:
      - description: 用户uuid
        in: body
        name: uuid
        required: true
        schema:
          type: string
      - description: 修改后的用户真实姓名
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: 用户头像url
        in: body
        name: avatar
        required: true
        schema:
          type: string
      - description: 用户个人简介描述信息
        in: body
        name: introduction
        required: true
        schema:
          type: string
      - description: 用户电话号码
        in: body
        name: phone
        required: true
        schema:
          type: string
      - description: 用户提币地址
        in: body
        name: address
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 更新用户信息失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 更新用户
      tags:
      - 用户管理
  /api/v1/wallet/actions:
    post:
      description: 根据用户uuid获取该用户的所有交易记录
      parameters:
      - description: 用户uuid
        in: body
        name: uuid
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 获取用户交易记录失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 交易记录
      tags:
      - 钱包管理
  /api/v1/wallet/info:
    post:
      description: 根据用户uuid获取用户的钱包数据
      parameters:
      - description: 用户uuid
        in: body
        name: uuid
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/app.Response'
        "500":
          description: 用户钱包数据失败
          schema:
            $ref: '#/definitions/app.Response'
      summary: 用户钱包数据
      tags:
      - 钱包管理
swagger: "2.0"